fix: resolved pr comments made by ikem
