+++ src/locales/en/common.json
@@ -270,7 +270,6 @@
   "Lisk Website": "Lisk Website",
   "Lisk accounts on {{WalletModel}}": "Lisk accounts on {{WalletModel}}",
   "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.": "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.",
-  "Lisk counts your message in bytes, so keep in mind\n            that the length of your message may vary in different languages.\n            Different characters may consume a varying amount of bytes.": "Lisk counts your message in bytes, so keep in mind\n            that the length of your message may vary in different languages.\n            Different characters may consume a varying amount of bytes.",
   "Lisk currently supports Ledger Nano S, Ledger Nano X wallets": "Lisk currently supports Ledger Nano S, Ledger Nano X wallets",
   "Lisk currently supports Ledger Nano S, Ledger Nano X, Trezor One and Trezor T wallets.": "Lisk currently supports Ledger Nano S, Ledger Nano X, Trezor One and Trezor T wallets.",
   "Lisk has now been enhanced even further": "Lisk has now been enhanced even further",
@@ -297,7 +296,6 @@
   "Members": "Members",
   "Message": "Message",
   "Message (Optional)": "Message (Optional)",
-  "Message (optional)": "Message (optional)",
   "Min": "Min",
   "Minimize": "Minimize",
   "Missed slot": "Missed slot",
@@ -590,7 +588,7 @@
   "Update to your new account": "Update to your new account",
   "Upload file is required": "Upload file is required",
   "Use maximum amount": "Use maximum amount",
-  "Use the sharing link to easily request any amount of tokens from Lisk Hub or Lisk Mobile users.": "Use the sharing link to easily request any amount of tokens from Lisk Hub or Lisk Mobile users.",
+  "Use the sharing link to easily request any amount of tokens from Lisk Desktop or Lisk Mobile users.": "Use the sharing link to easily request any amount of tokens from Lisk Desktop or Lisk Mobile users.",
   "Use your old passphrase ": "Use your old passphrase ",
   "Username can not be empty.": "Username can not be empty.",
   "Username is too long.": "Username is too long.",
