+++ src/modules/token/fungible/components/MessageField/MessageField.css
@@ -1,4 +1,4 @@
-@import '../../../../../setup/react/app/mixins.css';
+@import '../../../../../../setup/react/app/mixins.css';
 
 .container {
   align-items: flex-start;
