+++ src/modules/token/fungible/components/MessageField/MessageField.js
