+++ src/modules/token/fungible/components/MessageField/index.js
