+++ src/modules/wallet/context/dropdownContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const DropdownContext = createContext({ onChange: () => {}, selectedValue: null });
+
+export default DropdownContext;
